Avoid cyclic dependency
