+++ app/workflow_manager/models/__init__.py
@@ -4,8 +4,6 @@ from .workflow import Workflow
 from .payload import Payload
 from .workflow_run import WorkflowRun, LibraryAssociation
 from .library import Library
-from .state import State
-from .utils import WorkflowRunUtil, Status
 from .analysis import Analysis
 from .analysis_run import AnalysisRun
 from .analysis_context import AnalysisContext
